+++ examples/browser/wallet-web-extension/src/js/background.js
@@ -5,132 +5,68 @@ import '../img/icon-34.png'
 
 import { RpcWallet, Node, MemoryAccount, Crypto, BrowserRuntimeConnection } from 'AE_SDK_MODULES'
 
-// const account = MemoryAccount({
-//   keypair: {
-//     secretKey: 'YOUR_PRIV',
-//     publicKey: 'YOUR_PUB'
-//   }
-// })
-
-const account2 = MemoryAccount({ keypair: Crypto.generateKeyPair() })
-
-// Init accounts
-const accounts = [
-  // You can add your own account implementation,
-  // Account.compose({
-  //     init() {
-  //     },
-  //     methods: {
-  //         /**
-  //          * Sign data blob
-  //          * @function sign
-  //          * @instance
-  //          * @abstract
-  //          * @category async
-  //          * @rtype (data: String) => data: Promise[String]
-  //          * @param {String} data - Data blob to sign
-  //          * @return {String} Signed data blob
-  //          */
-  //         async sign(data) {
-  //         },
-  //         /**
-  //          * Obtain account address
-  //          * @function address
-  //          * @instance
-  //          * @abstract
-  //          * @category async
-  //          * @rtype () => address: Promise[String]
-  //          * @return {String} Public account address
-  //          */
-  //         async address() {
-  //         }
-  //     }
-  // })(),
-  // account,
-  account2
-]
-// Send wallet connection info to Aepp throug content script
+// ideally this can be configured by the users of the extension
 const NODE_URL = 'https://testnet.aeternity.io'
-const NODE_INTERNAL_URL = 'https://testnet.aeternity.io'
 const COMPILER_URL = 'https://compiler.aepps.com'
+const accounts = [
+  MemoryAccount({ keypair: Crypto.generateKeyPair() }), // generate keypair for account1
+  MemoryAccount({ keypair: Crypto.generateKeyPair() })  // generate keypair for account2
+]
 
 async function init () {
   // Init extension stamp from sdk
   RpcWallet({
     compilerUrl: COMPILER_URL,
-    nodes: [{ name: 'test-net', instance: await Node({ url: NODE_URL, internalUrl: NODE_INTERNAL_URL }) }],
+    nodes: [{ name: 'testnet', instance: await Node({ url: NODE_URL }) }],
     name: 'Wallet WebExtension',
-    // By default `ExtesionProvider` use first account as default account. You can change active account using `selectAccount (address)` function
+    // The `ExtensionProvider` uses the first account by default. You can change active account using `selectAccount(address)` function
     accounts,
     // Hook for sdk registration
     onConnection (aepp, action) {
-      // if (confirm(`Client ${aepp.info.name} with id ${aepp.id} want to connect`)) {
-      //   action.accept()
-      // } else {
-      //   action.deny()
-      // }
-      console.log('connection request')
-      console.log(aepp)
-      console.log(action)
-      console.log('------------------------')
-      action.accept()
+      if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} wants to connect`)) {
+        action.accept()
+      } else {
+        action.deny()
+      }
     },
     onDisconnect (msg, client) {
       console.log('Disconnect client: ', client)
     },
     onSubscription (aepp, action) {
-      // if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} want to subscribe for accounts`)) {
-      //   action.accept()
-      // } else {
-      //   action.deny()
-      // }
-      console.log('sign transaction')
-      console.log(aepp)
-      console.log(action)
-      console.log('------------------------')
-      action.accept()
+      if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} wants to subscribe for accounts`)) {
+        action.accept()
+      } else {
+        action.deny()
+      }
     },
     onSign (aepp, action) {
-      // if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} want to sign tx ${action.params.tx}`)) {
-      //   action.accept()
-      // } else {
-      //   action.deny()
-      // }
-      console.log('sign transaction')
-      console.log(aepp)
-      console.log(action)
-      console.log('------------------------')
-      action.accept()
+      if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} wants to sign tx ${action.params.tx}`)) {
+        action.accept()
+      } else {
+        action.deny()
+      }
     },
     onAskAccounts (aepp, action) {
-      // if (confirm(`Client ${aepp.info.name} with id ${aepp.id} want to get accounts`)) {
-      //   accept()
-      // } else {
-      //   deny()
-      // }
-      console.log(aepp)
-      console.log(action)
-      console.log('------------------------')
-      action.accept()
+      if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} wants to get accounts`)) {
+        action.accept()
+      } else {
+        action.deny()
+      }
     },
     onMessageSign (aepp, action) {
-      // if (confirm(`Client ${aepp.info.name} with id ${aepp.id} want to get accounts`)) {
-      //   accept()
-      // } else {
-      //   deny()
-      // }
-      console.log(aepp)
-      console.log(action)
-      console.log('------------------------')
-      action.accept()
+      if (confirm(`Aepp ${aepp.info.name} with id ${aepp.id} wants to sign msg ${action.params.message}`)) {
+        action.accept()
+      } else {
+        action.deny()
+      }
     }
   }).then(wallet => {
     chrome.runtime.onConnect.addListener(async function (port) {
-      // create Connection
+      // create connection
       const connection = await BrowserRuntimeConnection({ connectionInfo: { id: port.sender.frameId }, port })
       // add new aepp to wallet
       wallet.addRpcClient(connection)
-      // Share wallet details
+      // share wallet details
       wallet.shareWalletInfo(port.postMessage.bind(port))
       setTimeout(() => wallet.shareWalletInfo(port.postMessage.bind(port)), 3000)
     })
@@ -139,4 +75,4 @@ async function init () {
   })
 }
 
-init().then(_ => console.log('Started!'))
+init().then(_ => console.log('Wallet initialized!'))