+++ src/utils/swagger.js
@@ -46,8 +46,13 @@ export default async (specUrl, { spec, internalUrl, disableBigNumbers, disableCa
   spec = spec || await (await fetch(specUrl)).json()
   const jsonImp = disableBigNumbers ? JSON : JsonBig
 
-  if (['https://compiler.aepps.com', 'https://latest.compiler.aepps.com']
-    .includes(new URL(specUrl).origin)) {
+  if (
+    [
+      'https://compiler.aepps.com',
+      'https://latest.compiler.aepps.com'
+    ].includes(new URL(specUrl).origin) ||
+    new URL(specUrl).origin.includes('https://')
+  ) {
     spec.schemes = ['https']
   }