+++ test/integration/index.js
@@ -32,15 +32,18 @@ export const ignoreVersion = process.env.IGNORE_VERSION || false
 export const genesisAccount = MemoryAccount({ keypair: { publicKey, secretKey } })
 export const account = Crypto.generateKeyPair()
 
-export const BaseAe = async (params = {}) => Universal
-  .waitMined(true)
+export const BaseAe = async (params = {}, compose = {}) => Universal
   .compose({
-    deepProps: { Ae: { defaults: { interval: 50, attempts: 1200 } }, Swagger: { defaults: { debug: !!process.env.DEBUG } } }
-  })({
+    deepProps: { Ae: { defaults: { interval: 50, attempts: 1200 } } }
+  })
+  .compose(compose)({
     ...params,
     compilerUrl,
     ignoreVersion,
-    accounts: [...params.accounts || [], genesisAccount],
+    accounts: [
+      ...params.accounts || [],
+      ...params.withoutGenesisAccount ? [] : [genesisAccount]
+    ],
     nodes: [{ name: 'test', instance: await Node({ url, internalUrl, ignoreVersion }) }]
   })