+++ test/integration/rpc.js
@@ -17,12 +17,12 @@
 
 import { before, describe, it } from 'mocha'
 import { MemoryAccount, Node, RpcAepp, RpcWallet } from '../../src'
-import { buildTxHash, unpackTx } from '../../src/tx/builder'
+import { unpackTx } from '../../src/tx/builder'
 import { decode } from '../../src/tx/builder/helpers'
 import BrowserWindowMessageConnection from '../../src/utils/aepp-wallet-communication/connection/browser-window-message'
 import { getBrowserAPI, getHandler } from '../../src/utils/aepp-wallet-communication/helpers'
 import { METHODS, RPC_STATUS } from '../../src/utils/aepp-wallet-communication/schema'
-import { generateKeyPair, verify } from '../../src/utils/crypto'
+import { generateKeyPair, verify, hash } from '../../src/utils/crypto'
 import { compilerUrl, genesisAccount, internalUrl, networkId, publicKey, url, ignoreVersion } from './'
 
 describe('Aepp<->Wallet', function () {
@@ -273,11 +273,8 @@ describe('Aepp<->Wallet', function () {
       })
 
       const signedTx = await aepp.signTransaction(tx)
-      const { tx: { signatures: [signature] } } = unpackTx(signedTx)
-      const txWithNetwork = Buffer.concat([
-        Buffer.from(networkId),
-        buildTxHash(tx, { raw: true })
-      ])
+      const { tx: { signatures: [signature], encodedTx: { rlpEncoded } } } = unpackTx(signedTx)
+      const txWithNetwork = Buffer.concat([Buffer.from(networkId), hash(rlpEncoded)])
       const valid = verify(txWithNetwork, signature, decode(address))
       valid.should.be.equal(true)
     })