+++ test/unit/tx.js
@@ -19,7 +19,14 @@ import '../'
 import { describe, it } from 'mocha'
 import { encode as rlpEncode } from 'rlp'
 import { salt } from '../../src/utils/crypto'
-import { classify, commitmentHash, ensureNameValid, isNameValid, produceNameId } from '../../src/tx/builder/helpers'
+import {
+  classify,
+  commitmentHash,
+  ensureNameValid,
+  getMinimumNameFee,
+  isNameValid,
+  produceNameId
+} from '../../src/tx/builder/helpers'
 import BigNumber from 'bignumber.js'
 import { toBytes } from '../../src/utils/bytes'
 import { parseBigNumber } from '../../src/utils/bignumber'
@@ -77,6 +84,36 @@ describe('Tx', function () {
     })
 
     it('don\'t throws exception', () => ensureNameValid('asdasdasd.chain'))
+  })
+
+  describe('getMinimumNameFee', () => {
+    it('returns correct name fees', () => {
+      // based on: https://stackoverflow.com/questions/1349404/generate-random-string-characters-in-javascript
+      function randomName (length) {
+        let result = ''
+        const characters =
+          'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+        for (let i = 0; i < length; i++) {
+          result += characters.charAt(
+            Math.floor(Math.random() * characters.length)
+          )
+        }
+        return result
+      }
+
+      // protocol name fees: https://github.com/aeternity/protocol/blob/master/AENS.md#protocol-fees-and-protection-times
+      const nameFees = [
+        5702887, 3524578, 2178309, 1346269, 832040, 514229, 317811, 196418,
+        121393, 75025, 46368, 28657, 17711, 10946, 6765, 4181, 2584, 1597, 987,
+        610, 377, 233, 144, 89, 55, 34, 21, 13, 8, 5, 3
+      ]
+
+      for (let i = 0; i < nameFees.length; i++) {
+        getMinimumNameFee(randomName(i + 1) + '.chain')
+          .toString()
+          .should.be.equal(BigNumber(nameFees[i]).times(1e14).toString())
+      }
+    })
   })
 
   describe('isNameValid', () => {